bonus starting to work

+++ crss/printf/src/printf/print_text.c
@@ -6,7 +6,7 @@
 /*   By: ivmirand <ivmirand@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/10 14:00:35 by ivmirand          #+#    #+#             */
-/*   Updated: 2024/11/20 13:37:27 by ivmirand         ###   ########.fr       */
+/*   Updated: 2024/11/20 15:24:56 by ivmirand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ size_t	ft_print_address(void *addr, char *flags)
 		return (ft_print_string(NIL));
 	result = 0;
 	result += ft_print_string("0x");
-	result += ft_print_hex((unsigned long long)addr, FALSE, flags);
+	result += ft_print_hex(addr, FALSE, flags);
 	return (result);
 }
